+++ server.c
@@ -30,7 +30,7 @@ static void * serverThread(void * arg)
     return 0;
 }
 
-void startServer(int* initStatus, int* threadStop, uint64_t* num, int num_threads) {
+void startServer(int* initStatus, int* threadStop, uint64_t* num, long num_threads) {
     collatzInit(initStatus, &takeNextNum);
     if (initStatus) {
         collatzStart(threadStop, num, num_threads);
